requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
argon2-cffi=20.1.0=py37h7b6447c_1
astor=0.8.1=py37_0
attrs=20.1.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.1.5=py_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.7.22=0
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py37_0
cffi=1.14.2=py37he30daa8_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cupti=10.0.130=0
cycler=0.10.0=py37_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
expat=2.2.9=he6710b0_2
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.2=h5ab3b9f_0
gast=0.2.2=py37_0
glib=2.65.0=h3eb4bd4_0
gmp=6.2.0=he1b5a44_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37hf8bcb03_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=py37hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
icu=64.2=he1b5a44_1
importlib-metadata=1.7.0=py37_0
importlib_metadata=1.7.0=0
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.17.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py37_1
jupyter_client=6.1.6=py_0
jupyter_core=4.6.3=py37_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37hfd86e86_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.16=h516909a_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0=hd408876_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he6710b0_1
markdown=3.2.2=py37_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.3.1=1
matplotlib-base=3.3.1=py37h817c723_0
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.1.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
nbconvert=5.6.1=py37_1
nbformat=5.0.7=py_0
ncurses=6.2=he6710b0_1
notebook=6.1.1=py37_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
olefile=0.46=py37_0
openssl=1.1.1g=h7b6447c_0
opt_einsum=3.1.0=py_0
packaging=20.4=py_0
pandas=1.1.1=py37he6710b0_0
pandoc=2.10.1=0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
patsy=0.5.1=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py37_1
pickleshare=0.7.5=py37_1001
pillow=7.2.0=py37hb39fc2d_0
pip=20.2.2=py37_0
pixman=0.38.0=h516909a_1003
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.6=py_0
protobuf=3.13.0=py37hf484d3e_0
ptyprocess=0.6.0=py37_0
pycairo=1.19.1=py37h01af8b0_3
pycparser=2.20=py_2
pygments=2.6.1=py_0
pyparsing=2.4.7=py_0
pyrsistent=0.16.0=py37h7b6447c_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=py_0
python-igraph=0.8.2=py37hb78272c_0
python_abi=3.7=1_cp37m
pytz=2020.1=py_0
pyzmq=19.0.1=py37he6710b0_1
readline=8.0=h7b6447c_0
scipy=1.5.2=py37h0b6359f_0
seaborn=0.10.1=1
seaborn-base=0.10.1=py_1
send2trash=1.5.0=py37_0
setuptools=49.6.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
statsmodels=0.12.0=py37h8f50634_0
tensorboard=1.15.0=pyhb230dea_0
tensorflow=1.15.0=gpu_py37h0f0df58_0
tensorflow-base=1.15.0=gpu_py37h9dcbed7_0
tensorflow-estimator=1.15.1=pyh2649769_0
tensorflow-gpu=1.15.0=h0d30ee6_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
texttable=1.6.2=py_0
tk=8.6.10=hbc83047_0
tornado=6.0.4=py37h7b6447c_1
tqdm=4.48.2=py_0
traitlets=4.3.3=py37_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=0.16.1=py_0
wheel=0.35.1=py_0
wrapt=1.12.1=py37h7b6447c_1
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
zeromq=4.3.2=he6710b0_2
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
